Auth Module/requirements.txt
aiohttp==3.6.2
altgraph==0.10.2
apns2==0.7.1
appdirs==1.4.4
apt-clone==0.2.1
apturl==0.5.2
asgiref==3.2.10
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==20.2.0
autopep8==1.5.4
bdist-mpkg==0.5.0
beautifulsoup4==4.9.3
boto3==1.15.8
botocore==1.18.8
Brlapi==0.6.6
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
command-not-found==0.3
configobj==5.0.6
configparser==4.0.2
contextlib2==0.6.0.post1
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.1.1
cupshelpers==1.0
defer==1.0.6
distlib==0.3.1
Django==3.1.2
django-coreapi==2.3.0
django-cors-headers==3.5.0
django-openapi==0.1.1.10
django-push-notifications==2.0.0
django-rest-swagger==2.2.0
django-storages==1.10.1
djangorestframework==3.12.1
drf-yasg==1.17.1
enum34==1.1.6
filelock==3.0.12
future==0.17.1
google-api-core==1.22.3
google-api-python-client==1.12.3
google-auth==1.22.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
h2==2.6.2
hpack==3.0.0
http-ece==1.1.0
httplib2==0.18.1
hyper==0.7.0
hyperframe==3.2.0
idna==2.10
idna-ssl==1.1.0
importlib-metadata==2.0.0
importlib-resources==3.0.0
inflection==0.5.1
ipaddress==1.0.22
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
louis==3.5.0
macholib==1.5.1
Mako==1.0.7
MarkupSafe==1.1.1
modulegraph==0.10.4
multidict==4.7.6
netifaces==0.10.4
onboard==1.4.1
openapi-codec==1.3.2
packaging==20.4
pathlib2==2.3.5
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.13.0
psutil==5.4.2
py-vapid==1.7.1
py2app==0.7.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.6.0
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.1
pygobject==3.26.1
PyICU==1.9.8
PyJWT==1.7.1
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyparted==3.11.1
pyrsistent==0.17.3
python-apt==1.6.5+ubuntu0.3
python-dateutil==2.8.1
python-debian==0.1.32
python-xapp==1.8.1
python-xlib==0.20
pytz==2020.1
pywebpush==1.11.0
pyxdg==0.25
PyYAML==5.3.1
reportlab==3.4.0
requests==2.24.0
requests-unixsocket==0.1.5
rsa==4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
scandir==1.10.0
sessioninstaller==0.0.0
setproctitle==1.1.10
simplejson==3.17.2
singledispatch==3.4.0.3
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
swagger-spec-validator==2.7.3
system-service==0.3
systemd-python==234
toml==0.10.1
typing==3.7.4.1
typing-extensions==3.7.4.3
ubuntu-drivers-common==0.0.0
ufw==0.36
Unidecode==1.0.22
uritemplate==3.0.1
urllib3==1.25.10
virtualenv==20.0.31
xkit==0.0.0
xmltodict==0.12.0
yarl==1.6.0
zipp==3.2.0
